requirements.txt
groq==0.13.0
typing_extensions==4.12.2
fastapi==0.115.6
redis==5.2.0
pydantic==2.9.2
pydantic_core==2.23.4
gdeltdoc==1.5.0
python-dotenv==1.0.1
pandas==2.2.3
uvicorn==0.32.1
httpx==0.27.2